refactor: 리팩터링 방식 변경 1. 현재 문제점은 HandlerExecution에서 무조건 ModelAndView로 타입 캐스팅을 하여 레거시 컨트롤러가 HandlerExecution에서는 에러가 발생한다. 2. 그러면 ModelAndView에서 String, ModelAndView로 분기문을 나눌 수도 있지만, 이 경우 ModelAndView가 직접 JspView까지 만든다는 단점이 존재한다. 3. 그래서 차라리 HandlerExecution을 인터페이스로 변경하고, 전략패턴으로 ManualHandlerExecution과 AnnotationHandlerExecution으로 나눈다.
