+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/AnnotationHandlerExecution.java
@@ -1,0 +1,21 @@
+package webmvc.org.springframework.web.servlet.mvc.tobe;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.lang.reflect.Method;
+import webmvc.org.springframework.web.servlet.ModelAndView;
+
+public class AnnotationHandlerExecution implements HandlerExecution {
+
+    private final Object controller;
+    private final Method method;
+
+    public AnnotationHandlerExecution(final Object controller, final Method method) {
+        this.controller = controller;
+        this.method = method;
+    }
+
+    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
+        return (ModelAndView) method.invoke(controller, request, response);
+    }
+}
