+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/HandlerExecution.java
@@ -2,21 +2,9 @@ package webmvc.org.springframework.web.servlet.mvc.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.lang.reflect.Method;
 import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.JspView;
 
-public class HandlerExecution {
+public interface HandlerExecution {
 
-    private final Object controller;
-    private final Method method;
-
-    public HandlerExecution(final Object controller, final Method method) {
-        this.controller = controller;
-        this.method = method;
-    }
-
-    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
-        return (ModelAndView) method.invoke(controller, request, response);
-    }
+    ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) throws Exception;
 }
